W4-library-version11/server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
pymongo==4.6.0
flask-pymongo==2.3.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
